wIP, cleaned up code, need code review
